requirements.txt
# Dashboard Baker Flask - Dependências
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
python-dotenv==1.0.0
gspread==5.10.0
oauth2client==4.1.3
plotly==5.15.0
gunicorn==21.2.0
Werkzeug==2.3.7